+++ mteb/abstasks/Image/AbsTaskZeroShotClassification_.py
