+++ mteb/evaluation/evaluators/Image/ZeroShotClassificationEvaluator_.py
