+++ mteb/tasks/Image/ZeroShotClassification_/__init__.py
