+++ mteb/tasks/Image/ZeroShotClassification_/eng/RESISC45.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from mteb.abstasks.Image.AbsTaskZeroShotClassification import (
+from mteb.abstasks.Image.AbsTaskZeroShotClassification_ import (
     AbsTaskZeroShotClassification,
 )
 from mteb.abstasks.TaskMetadata import TaskMetadata
