+++ mteb/tasks/Image/ZeroShotClassification_/eng/templates/Country211_labels.txt
