+++ mteb/tasks/Image/ZeroShotClassification_/eng/templates/GTSRB_labels.txt
