+++ mteb/tasks/Image/ZeroShotClassification_/eng/templates/Imagenet1k_labels.txt
