+++ mteb/tasks/Image/ZeroShotClassification_/eng/templates/PatchCamelyon_labels.txt
