+++ mteb/tasks/Image/__init__.py
@@ -8,4 +8,4 @@ from .ImageClassification import *
 from .ImageMultilabelClassification import *
 from .ImageTextPairClassification import *
 from .VisualSTS import *
-from .ZeroShotClassification import *
+from .ZeroShotClassification_ import *
