+++ mteb/tasks/__init__.py
@@ -11,7 +11,7 @@ from .Image.ImageClassification import *
 from .Image.ImageMultilabelClassification import *
 from .Image.ImageTextPairClassification import *
 from .Image.VisualSTS import *
-from .Image.ZeroShotClassification import *
+from .Image.ZeroShotClassification_ import *
 from .InstructionRetrieval import *
 from .MultiLabelClassification import *
 from .PairClassification import *
